agent/requirements.txt
langchain-core
langgraph
langchain-deepseek
python-dotenv
